net: cnss2: Return 0 if PCI dev is not present in cnss_register_subsys
If plat_priv->pci_dev is NULL, the PCI device is not probed.
Earlier, in cnss_wlan_register_driver, if plat_priv->pci_dev is NULL,
cnss_register_subsys will not be called and CNSS2 would not try to
power on the target. Now driver invokes cnss_wlan_probe_driver which
misses this check.

In cnss_register_subsys, if plat_priv->pci_dev is NULL, -ENODEV is
returned to the driver which causes the entire driver load to fail.

Return 0 in cnss_register_subsys if plat_priv->pci_dev is NULL so that
driver skips this radio and continues to load for other radios.

Change-Id: I4e916c2b16f14220ab1583e6d26c02a896cbc1e7
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>