net: cnss2: Do not send QDSS Trace Mode message if FW is down
QDSS Trace Mode message can be triggered using cnsscli command. If FW is
down for a particular interface and cnsscli command is issued on that
interface, CNSS2 tries to send the QMI message to FW, fails as FW is
down and asserts.

Add check to send QDSS Trace Mode message only if FW_READY and
QMI_WLFW_CONNECTED states are set.

Change-Id: I007f5a35eef8e55d47a97a8158b1c7200ea1573b
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>
Signed-off-by: Hariharan Ramanathan <quic_hramanat@quicinc.com>