cnss2: Add a new QMI message to download FW INI CFG file
Add a new QMI message to send the contents of FW ini configuration
file which consists of information like the number of VAPs,
number of clients etc., for each memory mode.  This is done so as to
avoid adding new fields to the host capabilities message as new
parameters come in. With this approach, if there is a new parameter
that has to be added, it can be added to this file. This QMI message
is designed to be generic and can be used to download different files
by specifying the type of file. In this particular case, for backward
compatibility, add a new field in the host capabilities request QMI
message to indicate whether the FW CFG INI file is present or not.
When a different file has to be downloaded, the same infrastructure can
be used by specifying the file availability and type.

Change-Id: I84227a3b4b79860b6617d90d27e5fbc94f5d1347
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>