cnss2: Add new field to notify if REGDB is mandatory or not
* Update QMI headers to the latest.
* Add a new field to notify if REGDB download is mandatory or not.

Generally, if the regdb file is found, its contents would be sent
to target over QMI. The new field regdb_mandatory is added to the
target capabilities response message to help cnss2 driver decide
upon the action to be taken if the file is not found. If
regdb_mandatory is set to true by the target but the regdb file is
not found, then cnss2 driver will indicate an error and won't
proceed further. If it is set to false, but the regdb file is not
found, cnss2 driver will skip the download and proceed with the next
QMI message.

Change-Id: I16465f3955981b09ddd86bc64e1d0e725cf81429
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>