net: cnss2: Allocate meta_info segment for 4K to align with sysfs_read
meta_info segment is allocated to sizeof meta_info struct but ELF header
created as part of do_elf_ramdump aligns the bytes to 4KB. This causes a
slab out-of-bounds read when ramdump_read is called to read the
meta_info segment.

Allocate 4KB segment for meta_info to align it with the ELF header size.

Change-Id: I8bf0035956138b7c0800246a06e615d1583ae5b1
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>