cnss2: Fetch OTP fused board ID for QCN9224
Fetch OTP fused board ID for QCN9224 and store it in pci_priv.
This can be used in turn to identify whether the QCN9224 radio is
a single phy or a split phy radio and thus helpful in deciding
the firmware file to be downloaded in cases where DTS is not used.
Also, correct the end address of BAR WINDOW 0.

Change-Id: I87884242d1b59d134a0027b893fd4b41cc847f54
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>