qca-cnss: Change probe type to Sync when registering platform driver
For 2xPine, the compatible string is matching for both the Pines and
causing the probe order changes after a few reboot iterations.
This is because we are registering the platform driver as an Asynchronous
type and the VAPs are not coming up due to a mismatch between probe order
change and the wireless configuration.

Change probe type to Synchronous type when registering the platform driver
to avoid probe order change.

Change-Id: Ifa4a9f2e5e2eb456f39a60e10028904a56e1b945
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>