qca-cnss: Driver ops is NULL in down path, call BUG_ON for RDDM collection
Because the driver ops in the down path is NULL, it missed fatal callback
the from platform driver to the wifi driver and continuing the execution.
As a result of not getting the valid dump, all data is reset.

Call BUG_ON if driver ops are NULL instead of continuing execution when
recovery is disabled.

Change-Id: I9ba23916303a2a3a6cf31d5fadf0f489af94a388
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>