qca-cnss: Add remoteproc rootpd support
Add changes to register for rootpd device Miami.The
Miami target assert results in rootpd fatal notification first
before userpd notification which was not handled before because
cnss registers only for userpd susbystem.

This change is added to handle the rootpd fatal notification
from Miami, so it allows cnss to handle the fatal notification
and synchronize the dump collection and rproc start/stop sequence instead
of rproc boot sequence for AHB radio along with the PCI radio when
recovery is enabled.

Change-Id: I7add9820746672f7e25db21c34126c5911f0445e
Signed-off-by: Sham Muthayyan <quic_smuthayy@quicinc.com>