qca-cnss: Follow selective SSR recovery procedure for radios.
Move back to previous SSR recovery method for 11ax platforms
where recovery is taken care of by RPROC. Synchronous SSR recovery
was introduced to address MLO related requirements in WiFi Driver
which needs control to power-on the radios as part of recovery.
This mechanism does not scale well for IPQ5018 platform radios.
In the new method, if RPD (Root PD) crash is triggered in case of a UPD
(User PD) crash, the CNSS2 driver takes care of the RPD crash, but this
requires the RPD handle to be provided in the DTS file (done for IPQ53xx
and forward only). We can't partially handle UPD crash alone that arises
out of RPD crash. It has to either be completely handled by RPROC or
completely handled by CNSS2 driver. When the original crash is a UPD crash
and only UPD crash notification is sent to CNSS2, then both the methods
work fine. The problem arises only in the case of when there is an RPD
crash notification from RPROC. Thus, it is safe to stick to older SSR
recovery method for 11ax radios that came before IPQ53xx. This change
would fix the problem where SSR recovery doesn't work in IPQ5018 when we
instruct (through wifitool command) RPROC to trigger and notify RPD crash
for every UPD crash (default is UPD crash is only notified by RPROC when
there is a UPD crash and this works fine as such).

Change-Id: If0197310c0c37ebb9c8f4a38bb666b82c7a57972
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>