qca-cnss: reuse the memory for pci_priv
The memory of 'pci_priv' is malloced when PCI power up,
and it's not freed when PCI power down, though it will
be finally freed when PCI device destroyed, but it will
cause memory consumption accumulates if continuously do
'power up' and 'power down'.

So reuse the memory if it's already malloced to avoid
the potential issue.

Change-Id: I562c3f5fca9e39a2575f4670e3f0e5080a3a5349
Signed-off-by: Miaoqing Pan <quic_miaoqing@quicinc.com>