qca-cnss: Fetch Q6 timestamp only when Q6 is up
There is a pld call from wifi driver to cnss to get the timestamp of Q6
without checking Q6 is up or down. When Q6 is down and trying to read
the register, seeing panic.

Add the check to see whether Q6 is up or down. If Q6 is down, don't
allow to read the register and return simply

Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>
(cherry picked from commit 09b47e381eaa9e421b6d7070702920dc7475cf3e)

Change-Id: I438daa239287508782fad2f58f5adcf995edc61e