qca-cnss: Avoid PBL/SBL dump in Mission Mode
If PBL/SBL dump is triggered, it will continue to dump every 7 seconds
irrespective of state of the MHI environment. The function from timer
context tried to print the SBL log region which should not be accessed
when the device is in Mission Mode.

Add the check to skip the dump when the device is in Mission Mode.

Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>
Change-Id: Ic7e42f4c45723fd2190bef27061612feb119987c
(cherry picked from commit 3f0546de40e522391f88c7787ebda126358e77ce)