qca-cnss: set number of clients to FW always "1"
Setting num_clients = 2 in Host Cap request will be the issue when
cnss-daemon is not connected with CNSS2 and daemon-support is set.
Lets say if the cnss-daemon is supported and cnss-daemon is crashed
before setting the num_clients or cnss-daemon is not even started
for some reason, then the num_clients updated 2.

Sending num_clients = 2 to FW, the FW continuously waiting for the
second client to be connected and failed to send the FW Mem Indication
message to Host to proceed with the further QMI sequences and this
leads to FW ready timeout in APPS.

Set num_clients = 1 always to FW in Host Cap QMI message since the
cnss-daemon will always connects to the CNSS2 driver only.

Change-Id: I507f62baa7dd50fda43d4370540240b69d12fb1c
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>
(cherry picked from commit 6ff5225c065f31ec9ba91aae4506ac577b3cddfe)