qca-cnss: decouple the mlo configuration API to per group basis
In bootup or module reload case, the entire INI will be parsed and push
it to CNSS2 from Host driver via PLD and setup the MLO configuration
before Host Cap. QMI message.

As part of on-demand MLO configuration setup, the group ID will be the
argument for vendor command. Hence leverage the existing to support per
group basis.

Refactor the API to handle MLO configuration setup based on the group
ID.

Change-Id: Ie89805cd1d840e8bc40106b70211356cb6e06ad1
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>