qca-cnss: do immediate Host assert if RDDM event is received in down path
In down path we are calling CNSS_ASSERT for Host crash with dumping the
QMI history. There is a issue, when the driver is shutdown and the RDDM
event is received at the same time. When the RDDM event is received,
starting dumping the QMI history and MHI reset is happening in another
shutdown context and these leads to invalid dump due MHI reset.

Do immediate Host assert by calling BUG_ON in RDDM event handler if it
is driver shutdown flow. These will avoid the race between Host assert
and shutdown path.

Change-Id: Ia69bed54d9a73131a2456fac782413c2e9d0fb76
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>