qca-cnss: add the additional register dump and SRAM for PBL stuck
During boot-up, the platform driver loads the Secondary Boot Loader(SBL)
image, sets the boot address and size via the Boot Host Interface(BHI),
and waits for Q6 acknowledgment. However, the Q6 is stuck in the
Primary Boot Loader(PBL) and experiences a power-on reset failure.

To debug this issue, dumping the state of the Q6 register from the Host.
This register dump can help narrow down the root cause.

Change-Id: I4310825e57edaeafc038de419c900797d822f7a5
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>