qca-cnss: Reduce MHI buf_len and IPCR channel num_elements
MHI Allocates buffers of default size (64K) if buf_len is not defined in
the mhi_config. Then num_elements is used to allocate buffers of default
size for IPCR channel. This uses 64K * 32 = 2048K of memory.

IPCR channels is used for QMI messages which is used only during power
on sequence. QMI framework limits the max length of the buffer to 6K.

So to save memory, reduce the buffer len by setting mhi_config buf_len
to 8K and reduce the number of elements to 8. This will use 8K * 8 = 64K
of memory.

Change-Id: Ife93d2196bae2f6d92723aaa6f82994b7fdda99a
Signed-off-by: Raghavendran Madhavan <quic_raghmadh@quicinc.com>