qca-cnss: add the additional register dump for SBL stuck
During boot-up, the platform driver loads the Secondary Boot Loader(SBL)
image, sets the boot address and size via the Boot Host Interface(BHI).
However, the Q6 is stuck in the Secondary Boot Loader(SBL) and
experiences a power-on reset failure.

To debug this issue, dumping the state of the Q6 register from the Host.
This register dump can help narrow down the root cause.

Change-Id: I7068db96aaa33b4ce9de8e0f75461410bc46a3c3
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>