qca-cnss: Update MLO shared memory reset flow.
Adding changes to update MLO shared memory reset flow.
The Master chip in the mlo group(chip 0) is always
used as a reference to reset the MLO shared memory.

This is done because during QMI init shared memory is
reset and shared to firmware in first chip QMI transaction.

But, in WSI bypass scenario the master chip needs to be
updated since the master chip can also be bypassed.

Hence updating the masterchip id whenever the mlo config
is updated depending on the wsi remap state. If no chip is
bypassed then by default the master chip will be chip 0 in
MLO group. If the master chip is bypassed then the next chip
in the group will bceome the master chip.

Change-Id: I39fdd22033741786a8e4ae2fdae90bd46f14e039
Signed-off-by: Raghavendran Madhavan <quic_raghmadh@quicinc.com>