qca-cnss: increase the MHI SOC RESET timeout to 3sec
2.5sec is the WDOG bite timeout when the target is stuck or crashed.
However, the platform driver waits only 200ms and continues the shutdown
and bringup sequence of Q6 and this leads to failure in bringup.

Increase the MHI SOC RESET timeout from 200ms to 3000ms. This helps to
debug the issue if WDOD bite happens when the target is stuck or
crashed.

Change-Id: Iae533a960c862dabc549431b880b9a13ba3118f6
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>