qca-cnss: enable the MLO support if disabled in non-mission mode
MLO support is set for the Mission mode. When the user booting with
Factory Test Mode(FTM), MLO support is disabled properly. However, when
the user booting back to Mission mode, the MLO support is not enabled
and thus all the MLO VAPs are failed to create.

Enable back the MLO support properly if the support is already disabled
in non-mission mode.

Change-Id: I91cfc8b1cd8dd698e71ae978a294f7e488b65e4c
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>