qca-cnss: fix the NULL check for the QDSS free in down path
The auto variable is used for NULL check and its always getting updated
with the global variable and thus the auto variable is always true and
trying to unmap the address which is already freed.

To fix this issue, take the reference of the global variable and use
it for NULL check.
Free the QDSS memory when the QDSS feature is not enabled in kernel 6.6

Change-Id: I635182455101dde39deacf41e3ee44b6090f917b
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>