qca-cnss: mount the FW partition for PCI devices as well
This feature is supported only for low memory profile for memory
optimization. The FW mount call is happened only for AHB target and if
the AHB target is disabled, then the FW partition is not mounted and
hitting the host crash due to FW binaries are not present during the
fimrware download.

To fix this issue, move the cnss_mount_firmware function call to common
path where both the bus type getting called.

Change-Id: I18d0bbce478e972ecf3cee88cb04e7844bc0b9a6
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>