qca-cnss: collect the RDDM dump when kernel panic happens
Currently, the RDDM dump is not collected when the kernel panic and
target_asserted flag is set.

When the target is asserted, the dump will be collected in MHI status
notifier context but the kernel panic might come in different context
and disturbe the on going dump collection. But in kernel panic handler,
the assumption is its already collected in MHI staus notifier context
and skipping the dump collection again.

To fix this issue, call the cnss_bus_dev_crash_shutdown function always
and skip logic is already in place in cnss_bus_dev_crash_shutdown
function if dump is already collected.

Change-Id: I77461c4887e251151c42ba76f842372e6eba7883
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>
(cherry picked from commit 73ee9dd4ccf6e66c2918ee2accceb055cb098b28)