qca-cnss: add helper function to get the master SoC ID
For the MLO shared memory initization, the master device should
initialize the memory for REO. For the parallel bring up any device can
hit the initization path and thus have to wait for the mater device to
initialize first.

To fix this issue, add the helper function to get the master SoC ID.
This API can be used by host driver during the REO initialization.

Change-Id: I9d2334f8ea9c99358d8c049153a0d07722708b3f
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>