qca-cnss: add helper API for set/get static bypass support
Currently, when the static bypass is set, the Q6 will be bringup in
Mission mode and VAPs will be created in 11ax mode. In this case, the
power reduction is very low since the Q6 is up and running in non MLO
operation. Another way is, we can bring up the Q6 in 11ax mode and
trigger the external commands to put the Q6 down without Global SOC
reset. This will reduce the power consumption but the Q6 bring up first
and then down the Q6 after the complete bootup is not required and
affecting the boot KPI.

To add this support, add the helper APIs and export to Host driver. The host
driver will disable the parallel bring up if the static bypass is set.

Change-Id: Ib53abc0e52079f3184897cbfd706e5120c34f08d
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>