qca-cnss: fix static_bypass_support flag for non-MLO device
Currently, it returns true when MLO chip and non-MLO chip is present in
a RDP e.g: Miami+WKK+Pebble+York combination. In this, the York is
non-MLO device and returns true since its in plat_env_index loop.

To fix this issue, pass the plat_priv as argument instead of looping and
return true only when the MLO chips are bypassed. Always return false
for non-MLO devices.

Change-Id: Ia3ed33a8e383c56dc56673216dc0e1150a4d28da
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>