qca-cnss: Add support to trigger target assert using QMI
Currently we are using BUG_ON for asserting the Host and Target for
collecting the dump. This takes minor delay to trigger the target assert
and we missed the valid dump some time.

Using QMI message QMI_WLFW_SHUTDOWN_REQ_V01 to trigger crash for
both AHB and PCI radios. This will be used when WMI pipe is struck and
assert cannot be triggerred using WMI method.

Adding a Sysfs entry "trigger_qmi_crash" to exercise the QMI assert flow.

Change-Id: Ic22e5f30bed7863ac8dc00c85fee61b2f8dc8cd6
Signed-off-by: Raghavendran Madhavan <quic_raghmadh@quicinc.com>