qca-cnss: skip sending WSI disable/enable QMI message for AHB type
For the AHB type, the SSR sequence will be,
    WSI disable QMI message
    Q6 down
    Ramdump collection and
    Q6 up
    WSI enable QMI message
Since the ramdump collection is taking too much time to complete, the
WSI disable and enable window is high ~20sec.

Whereas, for PCI type the issue is not seen since the WSI
disable/enable happened after dump collection.
    Ramdump collection and
    WSI disable QMI message
    Q6 down
    Q6 up
    WSI enable QMI message

To address this issue for AHB, skip sending QMI message for AHB type.

Change-Id: I9a6b1f7a22196cc249f9c408d086e148a7add7e2
Signed-off-by: Madhavan Ganesan <madhgane@qti.qualcomm.com>